Move dv and pvc related code from virt controller vmi
