+++ pkg/storage/types/dv.go
@@ -31,12 +31,39 @@ import (
 	"kubevirt.io/client-go/kubecli"
 	"kubevirt.io/client-go/log"
 	cdiv1 "kubevirt.io/containerized-data-importer-api/pkg/apis/core/v1beta1"
+
 	"kubevirt.io/kubevirt/pkg/controller"
 )
 
 type CloneSource struct {
 	Namespace string
 	Name      string
+}
+
+func DataVolumeWFFC(dv *cdiv1.DataVolume) bool {
+	return dv.Status.Phase == cdiv1.WaitForFirstConsumer
+}
+
+func HasWFFCDataVolumes(dvs []*cdiv1.DataVolume) bool {
+	for _, dv := range dvs {
+		if DataVolumeWFFC(dv) {
+			return true
+		}
+	}
+	return false
+}
+
+func DataVolumeFailed(dv *cdiv1.DataVolume) bool {
+	return dv.Status.Phase == cdiv1.Failed
+}
+
+func HasFailedDataVolumes(dvs []*cdiv1.DataVolume) bool {
+	for _, dv := range dvs {
+		if DataVolumeFailed(dv) {
+			return true
+		}
+	}
+	return false
 }
 
 func GetCloneSourceWithInformers(vm *virtv1.VirtualMachine, dvSpec *cdiv1.DataVolumeSpec, dataSourceInformer cache.SharedIndexInformer) (*CloneSource, error) {
@@ -81,7 +108,7 @@ func GetCloneSourceWithInformers(vm *virtv1.VirtualMachine, dvSpec *cdiv1.DataVo
 	return cloneSource, nil
 }
 
-func GetCloneSource(ctx context.Context, client kubecli.KubevirtClient, vm *virtv1.VirtualMachine, dvSpec *cdiv1.DataVolumeSpec) (*CloneSource, error) {
+func GetCloneSource(ctx context.Context, client kubecli.KubevirtClient, namespace string, dvSpec *cdiv1.DataVolumeSpec) (*CloneSource, error) {
 	var cloneSource *CloneSource
 	if dvSpec.Source != nil && dvSpec.Source.PVC != nil {
 		cloneSource = &CloneSource{
@@ -90,10 +117,10 @@ func GetCloneSource(ctx context.Context, client kubecli.KubevirtClient, vm *virt
 		}
 
 		if cloneSource.Namespace == "" {
-			cloneSource.Namespace = vm.Namespace
+			cloneSource.Namespace = namespace
 		}
 	} else if dvSpec.SourceRef != nil && dvSpec.SourceRef.Kind == "DataSource" {
-		ns := vm.Namespace
+		ns := namespace
 		if dvSpec.SourceRef.Namespace != nil {
 			ns = *dvSpec.SourceRef.Namespace
 		}
@@ -118,6 +145,44 @@ func GetCloneSource(ctx context.Context, client kubecli.KubevirtClient, vm *virt
 	return cloneSource, nil
 }
 
+func GenerateDataVolumeFromTemplate(clientset kubecli.KubevirtClient, dataVolumeTemplate virtv1.DataVolumeTemplateSpec, namespace, priorityClassName string) (*cdiv1.DataVolume, error) {
+	newDataVolume := &cdiv1.DataVolume{}
+	newDataVolume.Spec = *dataVolumeTemplate.Spec.DeepCopy()
+	newDataVolume.ObjectMeta = *dataVolumeTemplate.ObjectMeta.DeepCopy()
+
+	labels := map[string]string{}
+	for k, v := range dataVolumeTemplate.Labels {
+		labels[k] = v
+	}
+	newDataVolume.ObjectMeta.Labels = labels
+
+	annotations := map[string]string{}
+	for k, v := range dataVolumeTemplate.Annotations {
+		annotations[k] = v
+	}
+	newDataVolume.ObjectMeta.Annotations = annotations
+
+	if newDataVolume.Spec.PriorityClassName == "" && priorityClassName != "" {
+		newDataVolume.Spec.PriorityClassName = priorityClassName
+	}
+
+	cloneSource, err := GetCloneSource(context.TODO(), clientset, namespace, &newDataVolume.Spec)
+	if err != nil {
+		return nil, err
+	}
+
+	if cloneSource != nil && newDataVolume.Spec.SourceRef != nil {
+		newDataVolume.Spec.SourceRef = nil
+		newDataVolume.Spec.Source = &cdiv1.DataVolumeSource{
+			PVC: &cdiv1.DataVolumeSourcePVC{
+				Namespace: cloneSource.Namespace,
+				Name:      cloneSource.Name,
+			},
+		}
+	}
+
+	return newDataVolume, nil
+}
 func GetDataVolumeFromCache(namespace, name string, dataVolumeInformer cache.SharedIndexInformer) (*cdiv1.DataVolume, error) {
 	key := controller.NamespacedKey(namespace, name)
 	obj, exists, err := dataVolumeInformer.GetStore().GetByKey(key)
@@ -152,7 +217,7 @@ func HasDataVolumeErrors(namespace string, volumes []virtv1.Volume, dataVolumeIn
 			continue
 		}
 
-		if dv.Status.Phase == cdiv1.Failed {
+		if DataVolumeFailed(dv) {
 			return fmt.Errorf("DataVolume %s is in Failed phase", volume.DataVolume.Name)
 		}
 
@@ -198,11 +263,7 @@ func HasDataVolumeProvisioning(namespace string, volumes []virtv1.Volume, dataVo
 }
 
 func ListDataVolumesFromTemplates(namespace string, dvTemplates []virtv1.DataVolumeTemplateSpec, dataVolumeInformer cache.SharedIndexInformer) ([]*cdiv1.DataVolume, error) {
-	var dataVolumes []*cdiv1.DataVolume
-
-	if len(dvTemplates) == 0 {
-		return dataVolumes, nil
-	}
+	dataVolumes := []*cdiv1.DataVolume{}
 
 	for _, template := range dvTemplates {
 		// get DataVolume from cache for each templated dataVolume
@@ -216,6 +277,60 @@ func ListDataVolumesFromTemplates(namespace string, dvTemplates []virtv1.DataVol
 		dataVolumes = append(dataVolumes, dv)
 	}
 	return dataVolumes, nil
+}
+
+func ListDataVolumesFromVolumes(namespace string, volumes []virtv1.Volume, dataVolumeInformer cache.SharedIndexInformer, pvcInformer cache.SharedIndexInformer) ([]*cdiv1.DataVolume, error) {
+	dataVolumes := []*cdiv1.DataVolume{}
+
+	for _, volume := range volumes {
+		dataVolumeName := getDataVolumeName(namespace, volume, pvcInformer)
+		if dataVolumeName == nil {
+			continue
+		}
+
+		dv, err := GetDataVolumeFromCache(namespace, *dataVolumeName, dataVolumeInformer)
+		if err != nil {
+			return dataVolumes, err
+		} else if dv == nil {
+			continue
+		}
+
+		dataVolumes = append(dataVolumes, dv)
+	}
+
+	return dataVolumes, nil
+}
+
+func getDataVolumeName(namespace string, volume virtv1.Volume, pvcInformer cache.SharedIndexInformer) *string {
+	if volume.VolumeSource.PersistentVolumeClaim != nil {
+		pvcInterface, pvcExists, _ := pvcInformer.GetStore().
+			GetByKey(fmt.Sprintf("%s/%s", namespace, volume.VolumeSource.PersistentVolumeClaim.ClaimName))
+		if pvcExists {
+			pvc := pvcInterface.(*v1.PersistentVolumeClaim)
+			pvcOwner := metav1.GetControllerOf(pvc)
+			if pvcOwner != nil && pvcOwner.Kind == "DataVolume" {
+				return &pvcOwner.Name
+			}
+		}
+	} else if volume.VolumeSource.DataVolume != nil {
+		return &volume.VolumeSource.DataVolume.Name
+	}
+	return nil
+}
+
+func DataVolumeByNameFunc(dataVolumeInformer cache.SharedIndexInformer, dataVolumes []*cdiv1.DataVolume) func(name string, namespace string) (*cdiv1.DataVolume, error) {
+	return func(name, namespace string) (*cdiv1.DataVolume, error) {
+		for _, dataVolume := range dataVolumes {
+			if dataVolume.Name == name && dataVolume.Namespace == namespace {
+				return dataVolume, nil
+			}
+		}
+		dv, exists, _ := dataVolumeInformer.GetStore().GetByKey(fmt.Sprintf("%s/%s", namespace, name))
+		if !exists {
+			return nil, fmt.Errorf("unable to find datavolume %s/%s", namespace, name)
+		}
+		return dv.(*cdiv1.DataVolume), nil
+	}
 }
 
 type DataVolumeConditionManager struct {
